Extract asynchronity utilities to separate file
