+++ lib/utilities.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const path = require('path');
-const { EventEmitter } = require('events');
 
 const random = (min, max) => {
   if (max === undefined) {
@@ -292,36 +291,6 @@ const parseCookies = (cookie) => {
   return values;
 };
 
-const createAbortController = () => {
-  const signal = new EventEmitter();
-  const abort = () => {
-    signal.emit('abort');
-  };
-  return { abort, signal };
-};
-
-const timeout = (msec, signal = null) =>
-  new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      reject(new Error('Timeout reached'));
-    }, msec);
-    if (!signal) return;
-    signal.on('abort', () => {
-      clearTimeout(timer);
-      reject(new Error('Timeout aborted'));
-    });
-  });
-
-const delay = (msec, signal = null) =>
-  new Promise((resolve, reject) => {
-    const timer = setTimeout(resolve, msec);
-    if (!signal) return;
-    signal.on('abort', () => {
-      clearTimeout(timer);
-      reject(new Error('Delay aborted'));
-    });
-  });
-
 module.exports = {
   random,
   sample,
@@ -347,7 +316,4 @@ module.exports = {
   makePrivate,
   protect,
   parseCookies,
-  createAbortController,
-  timeout,
-  delay,
 };
