+++ metautil.js
@@ -4,5 +4,6 @@ const utilities = require('./lib/utilities.js');
 const secutity = require('./lib/security.js');
 const semaphore = require('./lib/semaphore.js');
 const pool = require('./lib/pool.js');
+const async = require('./lib/async.js');
 
-module.exports = { ...utilities, ...secutity, ...semaphore, ...pool };
+module.exports = { ...utilities, ...secutity, ...semaphore, ...pool, ...async };
